+++ tests/test_delete.py
@@ -6,7 +6,7 @@ from scim2_models import User
 
 from scim2_client import RequestNetworkError
 from scim2_client import SCIMRequestError
-from scim2_client import SyncSCIMClient
+from scim2_client.engines.httpx import SyncSCIMClient
 
 
 def test_delete_user(httpserver):