+++ tests/test_utils.py
@@ -3,7 +3,7 @@ from scim2_models import Group
 from scim2_models import ServiceProviderConfig
 from scim2_models import User
 
-from scim2_client import SyncSCIMClient
+from scim2_client.engines.httpx import SyncSCIMClient
 
 
 def test_guess_resource_endpoint():